refactor view model to support base view model
