+++ app/src/main/java/com/berlin/aflami/di/viewModelModule.kt
@@ -1,7 +1,6 @@
 package com.berlin.aflami.di
 
 import com.berlin.aflami.viewmodel.searchactor.SearchByActorViewModel
-import com.berlin.aflami.viewmodel.mediadetails.MediaDetailsViewmodel
 import org.koin.core.module.dsl.viewModelOf
 import com.berlin.aflami.viewmodel.search.SearchViewModel
 import com.berlin.aflami.viewmodel.searchcountry.SearchByCountryViewModel
