+++ presentation/ui/src/main/kotlin/com/berlin/aflami/screens/mediadetails/components/CompanyProductionSection.kt
@@ -2,21 +2,18 @@ package com.berlin.aflami.screens.mediadetails.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.berlin.aflami.viewmodel.mediadetails.CompanyProductionItem
+import com.berlin.aflami.viewmodel.mediadetails.uistate.CompanyProductionUiState
 
 @Composable
 fun CompanyProductionSection(
-    companyProductions: List<CompanyProductionItem>,
+    companyProductions: List<CompanyProductionUiState>,
     columns: Int = 2 // Set this as needed (2 or 3 typical)
 ) {
     val rows = (companyProductions.size + columns - 1) / columns
