+++ presentation/ui/src/main/kotlin/com/berlin/aflami/screens/mediadetails/components/ReviewsSection.kt
@@ -4,12 +4,12 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
-import com.berlin.aflami.viewmodel.uistate.ReviewUiState
+import com.berlin.aflami.viewmodel.mediadetails.uistate.ReviewUiState
 
 @Composable
 fun ReviewsSection(
     reviews: List<ReviewUiState>,
-    isExpanded: Boolean,
+//    isExpanded: Boolean,
     onToggleExpand: () -> Unit
 ) {
     Column {
@@ -18,7 +18,7 @@ fun ReviewsSection(
             ReviewItem(
                 review = review,
                 isLastItem = isLast,
-                isExpanded = isExpanded,
+//                isExpanded = isExpanded,
                 onToggleExpand = onToggleExpand
             )
         }
