--- presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/mapper/MediaCast.kt
@@ -1,11 +1,0 @@
-package com.berlin.aflami.viewmodel.mapper
-
-import com.berlin.aflami.viewmodel.uistate.MediaCastUiState
-import com.berlin.entity.MediaCast
-
-fun MediaCast.toUiState():MediaCastUiState{
-    return MediaCastUiState(
-        name = name,
-        poster = poster
-    )
-}
