+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/mapper/Movie.kt
@@ -1,10 +1,8 @@
 package com.berlin.aflami.viewmodel.mapper
 
-import com.berlin.aflami.viewmodel.uistate.MediaUiState
-import com.berlin.aflami.viewmodel.uistate.MovieUIState
-import com.berlin.aflami.viewmodel.uistate.TVShowUiState
+import com.berlin.aflami.viewmodel.mediadetails.uistate.MediaUiState
+import com.berlin.aflami.viewmodel.shareduistate.MovieUIState
 import com.berlin.entity.Movie
-import com.berlin.entity.TVShow
 
 
 fun Movie.toUIStateMedia(): MediaUiState {
@@ -17,12 +15,12 @@ fun Movie.toUIStateMedia(): MediaUiState {
         poster = poster
     )
 }
-fun TVShow.toUIStateMedia(): MediaUiState {
-    return MediaUiState(
+fun Movie.toUIState(): MovieUIState {
+    return MovieUIState(
         id = id,
         title = title,
-        rating = rating.toString(),
-        releaseYear = releaseYear.toString(),
+        rating = rating.toString().take(3),
+        releaseYear = releaseYear.year.toString(),
         genre = genre,
         poster = poster
     )
