+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/mediadetails/uistate/MediaDetailsUiState.kt
@@ -1,21 +1,9 @@
-package com.berlin.aflami.viewmodel.uistate
+package com.berlin.aflami.viewmodel.mediadetails.uistate
 
 import androidx.compose.ui.graphics.painter.Painter
 import com.berlin.entity.Episodes
 
 enum class MediaType { MOVIE, TV_SHOW }
-
-data class MediaUiState(
-    val id: Long = 0L,
-    val title: String = "",
-    val rating: String = "",
-    val releaseYear: String = "",
-    val genre: List<Int> = emptyList(),
-    val poster: String = "",
-    val mediaType: String = "",
-
-    )
-
 
 data class MediaDetailsUiState(
     val id: Long = 0L,
@@ -39,9 +27,20 @@ data class MediaDetailsUiState(
     val isLoading: Boolean = true,
     val originalCountry: String? = null,
     val duration: String? = null,
-    val hasVideo: Boolean = false
+    val hasVideo: Boolean = false,
+    val error: String? = "",
+    val rowSection:RowSectionUiState = RowSectionUiState.Loading
 )
 
+data class MediaUiState(
+    val id: Long = 0L,
+    val title: String = "",
+    val rating: String = "",
+    val releaseYear: String = "",
+    val genre: List<Int> = emptyList(),
+    val poster: String = "",
+    val mediaType: String = ""
+)
 
 data class EpisodesUiState(
     val stillPath: String,
@@ -73,4 +72,21 @@ data class MediaCastUiState(
     val mediaId: Long = 0L,
     val name: String = "",
     val poster: String = "",
+)
+
+data class ReviewUiState(
+    val id: String = "",
+    val name: String = "",
+    val userName: String = "",
+    val avatarImage: String?,
+    val rating: Double = 0.0,
+    val content: String = "",
+    val date: String = "",
+)
+
+data class CompanyProductionUiState(
+    val id: String = "",
+    val image: String? = null,
+    val name: String = "",
+    val country: String = "",
 )
