--- presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/review/mapper.kt
@@ -1,16 +1,0 @@
-package com.berlin.aflami.viewmodel.review
-
-import com.berlin.aflami.viewmodel.uistate.ReviewUiState
-import com.berlin.entity.Review
-
-fun Review.toUiState(): ReviewUiState {
-    return ReviewUiState(
-        id = id,
-        name = name,
-        userName = userName,
-        avatarImage = avatarImage,
-        rating = rating,
-        content = content,
-        date = date.substringBefore("T")
-    )
-}
