+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/searchactor/SearchByActorScreenUiState.kt
@@ -1,7 +1,7 @@
 package com.berlin.aflami.viewmodel.searchactor
 
 import androidx.paging.PagingData
-import com.berlin.aflami.viewmodel.uistate.MediaUiState
+import com.berlin.aflami.viewmodel.mediadetails.uistate.MediaUiState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
