--- presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/uistate/ReviewState.kt
@@ -1,12 +1,0 @@
-package com.berlin.aflami.viewmodel.uistate
-
-import kotlin.String
-
-sealed class ReviewState {
-    sealed class Reviewing : ReviewState() {
-        object Loading : Reviewing()
-        data class Success(val data: List<ReviewUiState>) : Reviewing()
-        data class Error(val errorMessage: String) : Reviewing()
-    }
-    object NoReviewFound : ReviewState()
-}
