--- presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/uistate/ReviewUiState.kt
@@ -1,11 +1,0 @@
-package com.berlin.aflami.viewmodel.uistate
-
-data class ReviewUiState(
-    val id: String = "",
-    val name: String = "",
-    val userName: String = "",
-    val avatarImage: String?,
-    val rating: Double = 0.0,
-    val content: String = "",
-    val date: String = "",
-)
